# docs/openapi-campuscalm.yaml
openapi: 3.0.3
info:
  title: Campus Calm API
  version: 0.1.0
  description: API do MVP Campus Calm
paths:
  /api/access/requests/:
    get:
      operationId: access_requests_list
      tags:
      - Access Requests
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AccessRequest'
          description: ''
    post:
      operationId: access_requests_create
      tags:
      - Access Requests
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessRequestCreate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AccessRequestCreate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AccessRequestCreate'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequestCreate'
          description: ''
  /api/access/requests/{id}/:
    get:
      operationId: access_requests_detail
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Access Requests
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequest'
          description: ''
  /api/access/requests/{id}/approve/:
    post:
      operationId: access_requests_approve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Access Requests
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessDecision'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AccessDecision'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AccessDecision'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequest'
          description: ''
  /api/access/requests/{id}/reject/:
    post:
      operationId: access_requests_reject
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Access Requests
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessRequest'
          description: ''
  /api/access/triage/{id}/log/:
    get:
      operationId: access_requests_triage_log
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Access Requests
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AITriageLog'
          description: ''
        '404':
          description: Triage log not found.
...
  /api/pomodoro/start/:
    post:
      operationId: pomodoro_start
      tags:
      - Pomodoro
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PomodoroStart'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PomodoroStart'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PomodoroStart'
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSession'
          description: ''
  /api/pomodoro/stop/{id}/:
    post:
      operationId: pomodoro_stop
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
      tags:
      - Pomodoro
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PomodoroStop'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PomodoroStop'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PomodoroStop'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroSession'
          description: ''
  /api/pomodoro/summary/weekly/:
    get:
      operationId: pomodoro_weekly_summary
      tags:
      - Pomodoro
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PomodoroWeeklySummary'
          description: ''
...
components:
  schemas:
    AITriageLog:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        model_name:
          type: string
          readOnly: true
        input_payload:
          readOnly: true
        output_payload:
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - input_payload
      - model_name
      - output_payload
    AccessDecision:
      type: object
      properties:
        decided_plan:
          type: string
    AccessRequest:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        requester_email:
          type: string
          format: email
          maxLength: 254
        requester_name:
          type: string
          maxLength: 255
        requester_type:
          $ref: '#/components/schemas/RequesterTypeEnum'
        institution_name:
          type: string
          maxLength: 255
        estimated_users:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        wants_features: {}
        message:
          type: string
        status:
          $ref: '#/components/schemas/AccessRequestStatusEnum'
        recommended_plan:
          oneOf:
          - $ref: '#/components/schemas/PlanCodeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        decided_plan:
          oneOf:
          - $ref: '#/components/schemas/PlanCodeEnum'
          - $ref: '#/components/schemas/BlankEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        created_user:
          type: integer
          nullable: true
      required:
      - created_at
      - id
      - requester_email
      - requester_type
      - updated_at
    AccessRequestCreate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        requester_email:
          type: string
          format: email
          maxLength: 254
        requester_name:
          type: string
          maxLength: 255
        requester_type:
          $ref: '#/components/schemas/RequesterTypeEnum'
        institution_name:
          type: string
          maxLength: 255
        estimated_users:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
          nullable: true
        wants_features: {}
        message:
          type: string
        status:
          allOf:
          - $ref: '#/components/schemas/AccessRequestStatusEnum'
          readOnly: true
        recommended_plan:
          allOf:
          - $ref: '#/components/schemas/PlanCodeEnum'
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - recommended_plan
      - requester_email
      - requester_type
      - status
    AccessRequestStatusEnum:
      enum:
      - PENDING
      - APPROVED
      - REJECTED
      type: string
      description: |-
        * `PENDING` - Pending
        * `APPROVED` - Approved
        * `REJECTED` - Rejected
...
    BlankEnum:
      enum:
      - ''
...
    PlanCodeEnum:
      enum:
      - LITE
      - PRO
      - PILOT
      - STARTER
      - SCHOOL
      - ENTERPRISE
      type: string
      description: |-
        * `LITE` - Lite
        * `PRO` - Pro
        * `PILOT` - Piloto
        * `STARTER` - Starter
        * `SCHOOL` - School
        * `ENTERPRISE` - Enterprise
    PomodoroSession:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        focus_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        break_minutes:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          format: int64
        started_at:
          type: string
          format: date-time
          readOnly: true
        ended_at:
          type: string
          format: date-time
          readOnly: true
          nullable: true
        status:
          allOf:
          - $ref: '#/components/schemas/PomodoroStatusEnum'
          readOnly: true
      required:
      - ended_at
      - id
      - started_at
      - status
    PomodoroStart:
      type: object
      properties:
        focus_minutes:
          type: integer
          minimum: 1
          default: 25
        break_minutes:
          type: integer
          minimum: 1
          default: 5
    PomodoroStatusEnum:
      enum:
      - COMPLETED
      - STOPPED
      type: string
      description: |-
        * `COMPLETED` - Completed
        * `STOPPED` - Stopped
    PomodoroStop:
      type: object
      properties:
        completed:
          type: boolean
          default: false
    PomodoroWeeklySummary:
      type: object
      properties:
        total_sessions:
          type: integer
        total_focus_minutes:
          type: integer
      required:
      - total_focus_minutes
      - total_sessions
...
    RequesterTypeEnum:
      enum:
      - INDIVIDUAL
      - INSTITUTION
      type: string
      description: |-
        * `INDIVIDUAL` - Individual
        * `INSTITUTION` - Institution
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT